+++ src/buffer.h
@@ -5,6 +5,7 @@
 #include <stdexcept>
 #include <string.h>
 #include <vector>
+#include <sstream>
 
 #if !defined(__arm__) && !defined(__aarch64__)
     #define BUFFER_ALLOW_UNALIGNED_MEMORY_ACCESS 1
@@ -95,7 +96,7 @@ public:
     {
         assert((mDataSize-offset) % sizeof(T) == 0);
         size_t count = (mDataSize-offset)/sizeof(T);
-        read(offset, output, count);
+        read(offset, output, static_cast<int>(count));
     }
     template <class T>
     void read(size_t offset, T& output)
